requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
trafilatura==1.12.2
lancedb==0.17.0
pyarrow==18.1.0
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.51.0
orjson==3.10.7
